+++ src/module/rolls/preroll-dialog.ts
@@ -16,6 +16,7 @@ import { renderRollGraphic } from './roll-graphic'
 import { CharacterDataProperties } from '../actor/actortypes'
 import { IronswornRollMessage } from '.'
 import { formatRollPlusStat } from './ironsworn-roll-message.js'
+import { ChallengeResolutionDialog } from './challenge-resolution-dialog'
 
 export function localeCapitalize(str: string) {
   const locale = game.i18n.lang
@@ -335,14 +336,20 @@ export class IronswornPrerollDialog extends Dialog<
     }).render(true)
   }
 
-  private static submitRoll(
+  private static async submitRoll(
     el: HTMLElement | JQuery<HTMLElement>,
     opts: PreRollOptions
   ) {
     const realOpts = prerollOptionsWithFormData($(el).find('form'), opts)
 
     const r = new IronswornRoll(realOpts)
-    return new IronswornRollMessage(r).createOrUpdate()
+    const msg = new IronswornRollMessage(r)
+    await msg.createOrUpdate()
+
+    // Show resolution dialog if needed
+    if (r.preRollOptions.extraChallengeDice) {
+      ChallengeResolutionDialog.showForMessage(msg.roll.chatMessageId ?? '')
+    }
   }
 
   private static async renderContent(data: {
@@ -353,15 +360,12 @@ export class IronswornPrerollDialog extends Dialog<
     showActorSelect?: boolean
     action?: boolean
   }): Promise<string> {
-    const newOptions = {
-      ...data.prerollOptions,
-      showOutcome: true,
-    }
-    console.log('preroll render options', newOptions)
-    const graphic = await renderRollGraphic(newOptions)
+    const graphic = await renderRollGraphic({
+      preRollOptions: data.prerollOptions,
+    })
     const template =
       'systems/foundry-ironsworn/templates/rolls/preroll-dialog.hbs'
-    return renderTemplate(template, { ...data, graphic, showOutcome: true })
+    return renderTemplate(template, { ...data, graphic })
   }
 
   activateListeners(html: JQuery<HTMLElement>): void {
@@ -380,7 +384,7 @@ export class IronswornPrerollDialog extends Dialog<
         this.element.find('form'),
         this.prerollOptions
       )
-      const graphic = await renderRollGraphic(pro)
+      const graphic = await renderRollGraphic({ preRollOptions: pro })
       this.element.find('.roll-graphic').replaceWith(graphic)
     }
     html.find('input').on('change', rerender)